transform-jao2eds/requirements.txt
apache-beam
pandas
osiris-sdk